+++ NEWS.md
@@ -6,9 +6,12 @@ New language features
 
 * Structs with all isbits and isbitsunion fields are now stored inline in arrays ([#32448]).
 
+* `import` now allows quoted symbols, e.g. `import Base.:+` ([#33158]).
+
 Language changes
 ----------------
 
+* Calling `show` or `repr` on an `undef`/`UndefInitializer()` array initializer now shows valid Julia code ([#33211]).
 
 Multi-threading changes
 -----------------------
@@ -25,6 +28,7 @@ New library functions
 * The `tempname` function now takes an optional `parent::AbstractString` argument to give it a directory in which to attempt to produce a temporary path name ([#33090]).
 * The `tempname` function now takes a `cleanup::Bool` keyword argument defaulting to `true`, which causes the process to try to ensure that any file or directory at the path returned by `tempname` is deleted upon process exit ([#33090]).
 * The `readdir` function now takes a `join::Bool` keyword argument defaulting to `false`, which when set causes `readdir` to join its directory argument with each listed name ([#33113]).
+* The new `only(x)` function returns the one-and-only element of a collection `x`, and throws an `ArgumentError` if `x` contains zero or multiple elements. ([#33129])
 
 
 Standard library changes
@@ -32,8 +36,9 @@ Standard library changes
 
 * The methods of `mktemp` and `mktempdir` which take a function body to pass temporary paths to no longer throw errors if the path is already deleted when the function body returns ([#33091]).
 
-#### Libdl
+* Verbose `display` of `Char` (`text/plain` output) now shows the codepoint value in standard-conforming `"U+XXXX"` format ([#33291]).
 
+#### Libdl
 
 #### LinearAlgebra
 
@@ -43,7 +48,11 @@ Standard library changes
 
 #### Random
 
+
 * `AbstractRNG`s now behave like scalars when used in broadcasting ([#33213]).
+
+* Products involving sparse arrays now allow more general sparse `eltype`s, such as `StaticArrays` ([#33205])
+
 
 #### SparseArrays